# packages/ecommerce-server/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"DGLResponse_any_": {
				"properties": {
					"data": {},
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"apiKey": {
						"type": "string"
					},
					"datasource": {
						"items": {
							"$ref": "#/components/schemas/Datasource"
						},
						"type": "array"
					},
					"isActive": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"username",
					"apiKey",
					"datasource",
					"isActive",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Datasource": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"platform": {
						"type": "string",
						"enum": [
							"Woocommerce",
							"Magento"
						]
					},
					"baseUrl": {
						"type": "string"
					},
					"consumerKey": {
						"type": "string"
					},
					"consumerSecret": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"user",
					"platform",
					"baseUrl",
					"consumerKey",
					"consumerSecret",
					"isActive",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DGLResponse_User_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/User"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUserReq": {
				"properties": {
					"username": {
						"type": "string"
					},
					"apiKey": {
						"type": "string"
					},
					"masterKey": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"apiKey",
					"masterKey"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DGLResponse_Datasource_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/Datasource"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUserDatasourceReq": {
				"properties": {
					"id": {
						"type": "string"
					},
					"platform": {
						"type": "string",
						"enum": [
							"Woocommerce",
							"Magento"
						]
					},
					"consumerKey": {
						"type": "string"
					},
					"consumerSecret": {
						"type": "string"
					},
					"baseUrl": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"platform",
					"consumerKey",
					"consumerSecret",
					"baseUrl"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "@dg-live/ecommerce-server",
		"version": "0.0.1",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/catalog": {
			"get": {
				"operationId": "GetCatalog",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DGLResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"Catalog"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "apiKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "datasourceId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/catalog/sync": {
			"get": {
				"operationId": "SyncCatalog",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DGLResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"Catalog"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "apiKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "datasourceId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/dashboard/user": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DGLResponse_User_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveUserReq"
							}
						}
					}
				}
			}
		},
		"/dashboard/user/keys": {
			"post": {
				"operationId": "SaveUserKeys",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DGLResponse_Datasource_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "apiKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveUserDatasourceReq"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetUserDatasources",
				"responses": {
					"200": {
						"description": "User",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DGLResponse_User_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "apiKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/v1"
		}
	]
}